requirements.txt
lxml==4.6.3
numpy==1.21.1
openpyxl==3.0.7
pandas==1.3.1
pyecharts @ file:./package/pyecharts-1.9.0
requests==2.26.0
retrying==1.3.3
simplejson==3.17.3
TA-Lib @ file:./package/TA_Lib-0.4.21-cp37-cp37m-win_amd64.whl
Flask==1.1.4
Flask-Caching==1.10.1
Flask-Migrate==3.1.0
Flask-MySQL==1.5.2
Flask-Script==2.0.6
Flask-WTF==0.15.1
jwt==1.2.0
jsonschema==3.2.0
mysqlclient==2.0.3
psycopg2-binary==2.8.6
pycryptodome==3.9.9
redis==3.5.3
SQLAlchemy==1.4.22
uWSGI
nacos-sdk-python==0.1.5
